added basic scenario getting quote

+++ src/main/scala/forex/domain/Price.scala
@@ -3,6 +3,6 @@ package forex.domain
 case class Price(value: BigDecimal) extends AnyVal
 
 object Price {
-  def apply(value: Integer): Price =
+  def apply(value: Double): Price =
     Price(BigDecimal(value))
 }
